fix: add output-shell command
