+++ pkg/controller/setshell/zsh.sh
@@ -1,0 +1,5 @@
+_aqua_env() {
+    eval "$(aqua output-shell)"
+}
+autoload -Uz add-zsh-hook
+add-zsh-hook preexec _aqua_env
